Adding BFD set bonuses
